+++ src/components/DialogWizard/DialogWizard.js
@@ -22,6 +22,7 @@ const DialogWizard = ({
   nextText,
   cancelText,
   disabled,
+  banner,
   ...otherProps
 }) => {
   const [step, setStep] = useState(currentStep || 1)
@@ -48,9 +49,9 @@ const DialogWizard = ({
       <div className={styles.steps}>
         <small>{languageService.getTranslation('step')}</small>
         <div className={styles.stepsNumbers}>
-          <small>{step}</small>
-          <small>-</small>
-          <small>{steps.length}</small>
+          <small className={styles.stepNumber}>{step}</small>
+          <small>{languageService.getTranslation('stepOutOf')}</small>
+          <small className={styles.stepNumber}>{steps.length}</small>
         </div>
       </div>
     )
@@ -110,6 +111,7 @@ const DialogWizard = ({
           style={{ width: `${(step / steps.length) * 100}%` }}
         />
       </div>
+      {banner}
       <div className={styles.content}>
         <Dialog.Body>{steps.length && steps[step - 1]}</Dialog.Body>
         <Dialog.Footer>
@@ -161,7 +163,7 @@ DialogWizard.propTypes = {
   disableBackdropClick: propTypes.bool,
   /** Disable onClose firing when escape button is clicked */
   disableEscapeKeyDown: propTypes.bool,
-  /** footer meesage can be a string or an element */
+  /** footer message can be a string or an element */
   footerMessage: propTypes.oneOfType([propTypes.string, propTypes.element]),
   /** header title can be a string or an element */
   headerTitle: propTypes.oneOfType([propTypes.string, propTypes.element]),
@@ -171,7 +173,7 @@ DialogWizard.propTypes = {
   currentStep: propTypes.number,
   /** overlay content element*/
   overlayContent: propTypes.element,
-  /** css costumization for ovarlay */
+  /** css customization for ovarlay */
   overlayClassName: propTypes.string,
   /** text for done button */
   finishText: propTypes.string,
@@ -181,6 +183,8 @@ DialogWizard.propTypes = {
   cancelText: propTypes.string,
   /** boolean to disable next/finish button */
   disabled: propTypes.bool,
+  /** banner component to be rendered */
+  banner: propTypes.element,
 }
 
 export default DialogWizard